refactor: GameService 분리
