+++ com.spoticket.games/src/main/java/com/spoticket/game/application/service/GameCommandService.java
@@ -5,26 +5,22 @@ import com.spoticket.game.domain.repository.GameRepository;
 import com.spoticket.game.dto.request.CreateGameRequest;
 import com.spoticket.game.dto.request.UpdateGameRequest;
 import com.spoticket.game.dto.response.GameResponse;
-import com.spoticket.game.dto.response.GenericPagedModel;
 import com.spoticket.game.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.UUID;
 
-import static org.springframework.http.HttpStatus.*;
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class GameService {
+@Transactional
+public class GameCommandService {
 
     private final GameRepository gameRepository;
 
-    @Transactional
     public GameResponse createGame(CreateGameRequest request) {
         Game game = Game.of(
                 request.getTitle(), request.getStartTime(), request.getSport(), request.getLeague(),
@@ -33,25 +29,6 @@ public class GameService {
         return GameResponse.from(gameRepository.save(game));
     }
 
-    public GameResponse getGame(UUID gameId) throws CustomException {
-        Game game = gameRepository.findByGameIdAndIsDeleteFalse(gameId)
-                .orElseThrow(() -> new CustomException(NOT_FOUND));
-        return GameResponse.from(game);
-    }
-
-    @Transactional
-    public void deleteGame(UUID gameId) {
-        Game game = gameRepository.findById(gameId)
-                .orElseThrow(() -> new CustomException(NOT_FOUND));
-        game.delete();
-    }
-
-    public GenericPagedModel<GameResponse> getGamesByStadiumId(UUID stadiumId, Pageable pageable) {
-        Page<GameResponse> page = gameRepository.findAllByStadiumIdAndIsDeleteFalse(stadiumId, pageable);
-        return GenericPagedModel.of(page);
-    }
-
-    @Transactional
     public GameResponse updateGame(UUID gameId, UpdateGameRequest request) {
         Game game = gameRepository.findById(gameId)
                 .orElseThrow(() -> new CustomException(NOT_FOUND));
@@ -67,5 +44,10 @@ public class GameService {
         return GameResponse.from(game);
     }
 
-}
+    public void deleteGame(UUID gameId) {
+        Game game = gameRepository.findById(gameId)
+                .orElseThrow(() -> new CustomException(NOT_FOUND));
+        game.delete();
+    }
 
+}
