+++ com.spoticket.games/src/main/java/com/spoticket/game/presentation/controller/GameController.java
@@ -1,6 +1,7 @@
 package com.spoticket.game.presentation.controller;
 
-import com.spoticket.game.application.service.GameService;
+import com.spoticket.game.application.service.GameCommandService;
+import com.spoticket.game.application.service.GameQueryService;
 import com.spoticket.game.dto.request.CreateGameRequest;
 import com.spoticket.game.dto.request.UpdateGameRequest;
 import com.spoticket.game.dto.response.GameResponse;
@@ -22,7 +23,8 @@ import static com.spoticket.game.global.util.ResponseUtils.*;
 @RequiredArgsConstructor
 public class GameController {
 
-    private final GameService gameService;
+    private final GameCommandService gameCommandService;
+    private final GameQueryService gameQueryService;
 
     @PostMapping
     public DataResponse<GameResponse> createGame(
@@ -30,27 +32,27 @@ public class GameController {
             BindingResult result
     ) throws CustomException {
         validateAndThrowIfInvalid(result);
-        return created(gameService.createGame(request));
+        return created(gameCommandService.createGame(request));
     }
 
     @GetMapping("/{gameId}")
     public DataResponse<GameResponse> getGame(@PathVariable UUID gameId) {
-        return ok(gameService.getGame(gameId));
+        return ok(gameQueryService.getGame(gameId));
     }
 
     @GetMapping
     public DataResponse<PagedModel<GameResponse>> getGamesByStadiumId(UUID stadiumId, Pageable pageable) {
-        return ok(gameService.getGamesByStadiumId(stadiumId, pageable));
+        return ok(gameQueryService.getGamesByStadiumId(stadiumId, pageable));
     }
 
     @PatchMapping("/{gameId}")
     public DataResponse<GameResponse> updateGame(@PathVariable UUID gameId, @RequestBody UpdateGameRequest request) {
-        return ok(gameService.updateGame(gameId, request));
+        return ok(gameCommandService.updateGame(gameId, request));
     }
 
     @DeleteMapping("/{gameId}")
     public BasicResponse deleteGame(@PathVariable UUID gameId) {
-        gameService.deleteGame(gameId);
+        gameCommandService.deleteGame(gameId);
         return noContent();
     }
